refactored to align more with SOLID principles, further split into modules for cleaner code
