+++ src/scripts/modules/menu.js
@@ -1,24 +1,3 @@
-
-export const menuItems = [
-  { name: 'Chicken & Waffles', price: '$12.99' },
-  { name: 'Biscuits & Gravy', price: '$9.99' },
-  { name: 'Pancakes', price: '$7.99' },
-  { name: 'French Toast', price: '$7.99' },
-  { name: 'Eggs Benedict', price: '$10.99' },
-  { name: 'Omelette', price: '$9.99' },
-  { name: 'Breakfast Burrito', price: '$8.99' },
-  { name: 'Breakfast Sandwich', price: '$7.99' },
-  { name: 'Bacon', price: '$3.99' },
-  { name: 'Sausage', price: '$3.99' },
-  { name: 'Hash Browns', price: '$2.99' },
-  { name: 'Home Fries', price: '$2.99' },
-  { name: 'Toast', price: '$1.99' },
-  { name: 'Coffee', price: '$1.99' },
-  { name: 'Tea', price: '$1.99' },
-  { name: 'Orange Juice', price: '$1.99' },
-  { name: 'Milk', price: '$1.99' }
-];
-
 const buildMenuList = (menuItemsArr) => {
   const menuList = document.createElement('ul');
